requirements.txt
django
python-decouple
google-generativeai
sentence-transformers
scikit-learn
numpy
PyPDF2
gunicorn
whitenoise
